requirements.txt
requests
beautifulsoup4
lxml
pandas
numpy>=1.24.0 
scikit-learn==1.4.1
plotly
streamlit
python-dotenv
trafilatura

# Dependencies for Internal Semantic Search
langchain
chromadb

# Dependencies for Advanced Content Extraction (Headless Browser)
selenium
webdriver-manager